Soil moisture shows as a percentage clamping values from 0 to 880. Bluetooth page no longer has history selected.

+++ app/src/main/java/com/themakers/plantlink/Bluetooth/ConnectedThread.kt
@@ -63,7 +63,7 @@ class ConnectedThread(_socket: BluetoothSocket, private var plantViewModel: Plan
                     } else if (readingNumber == 2) {
                         plantViewModel.setHumid(readMessage.toDouble())
                     } else if (readingNumber == 3) {
-                        plantViewModel.setMoist(readMessage.toInt())
+                        plantViewModel.setMoist(readMessage.toDouble())
                     }
 
                     bytes = 0
